# docs/swagger.yaml
basePath: /
definitions:
  admin.CreateTenantResult:
    properties:
      owner:
        $ref: '#/definitions/admin.OwnerResult'
      tenant:
        $ref: '#/definitions/admin.TenantResult'
    type: object
...
  admin.OwnerResult:
    properties:
      email:
        type: string
      id:
        type: string
      invite_error:
        type: string
      invite_sent:
        type: boolean
      role:
        type: string
      status:
        type: string
    type: object
  admin.PermissionResult:
    properties:
      application_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      tenant_id:
        type: string
    type: object
...
  admin.TenantResult:
    properties:
      cors_origins:
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      domain:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      plan:
        type: string
      region:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
  admin.TenantsPage:
    properties:
      data:
        items:
          $ref: '#/definitions/admin.TenantResult'
        type: array
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handlers.CreatePermissionRequest:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
  handlers.CreateTenantRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      domain:
        type: string
      name:
        type: string
      owner_email:
        type: string
      plan:
        type: string
      region:
        type: string
      slug:
        type: string
    type: object
...
  handlers.UpdateTenantRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      domain:
        type: string
      name:
        type: string
      plan:
        type: string
      region:
        type: string
    type: object
...
host: localhost:8082
info:
  contact:
    name: EngineersMind
    url: https://github.com/engineersmind/emc-auth-server
  description: Standalone multi-tenant Identity Provider — email/password auth, JWT,
    refresh tokens, TOTP 2FA, SAML 2.0, RBAC, Admin UI, and AI/Agent security.
  license:
    name: MIT
  title: EMC Auth Server
  version: 1.5.0
paths:
...
  /api/v1/permissions/{id}:
    delete:
      description: Deletes a permission from the caller's tenant. Cascades to roles
        and users. Requires admin:access.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete permission
      tags:
      - admin-rbac
    put:
      consumes:
      - application/json
      description: Renames a permission and/or replaces its description. Empty name
        keeps the current one. Requires admin:access.
      parameters:
      - description: Permission ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.PermissionResult'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Permission name already exists
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update permission
      tags:
      - admin-rbac
...
  /api/v1/tenants:
    get:
      description: 'Returns tenants scoped to the caller''s permissions: callers with
        tenant:manage get a paginated, filtered list of every tenant; any other authenticated
        caller gets only the tenants tied to their own account email, each with their
        role and usage stats (search/status/region/pagination params are ignored in
        that case).'
      parameters:
      - description: Search by name, display name, or domain
        in: query
        name: search
        type: string
      - description: 'Filter by status: active | inactive | suspended (suspended maps
          to inactive)'
        in: query
        name: status
        type: string
      - description: Filter by region (exact match)
        in: query
        name: region
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: 'Rows per page (default 25, max 100); alias: limit'
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.TenantsPage'
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List tenants
      tags:
      - admin-tenants
    post:
      consumes:
      - application/json
      description: 'Creates a new isolated tenant and auto-seeds the granular permission
        catalog, an owner role holding all of it, an owner user for the provided owner_email,
        and the matching tenant_admins record. The owner account is created WITHOUT
        a password: an invitation link is emailed, and accepting it sets the password
        and marks the address verified. The response carries owner.status = "pending_invitation"
        and owner.invite_sent; when invite_sent is false the tenant still exists and
        the invitation can be resent. Requires tenant:manage permission (super_admin
        only).'
      parameters:
      - description: Tenant details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/admin.CreateTenantResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Slug already taken
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create tenant
      tags:
      - admin-tenants
  /api/v1/tenants/{id}:
    delete:
      description: Soft-deactivates a tenant (sets is_active=false). Requires tenant:manage
        permission.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Deactivate tenant
      tags:
      - admin-tenants
    get:
      description: Returns a single tenant by ID. Requires tenant:manage permission.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.TenantResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tenant
      tags:
      - admin-tenants
    put:
      consumes:
      - application/json
      description: Updates editable fields on a tenant. Requires tenant:manage permission.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin.TenantResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update tenant
      tags:
      - admin-tenants
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer <your_access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"